LOCAL_MSG_OPTIM: add groups and fixed for anytime migration
LOCAL_MSG_OPTIM avoids copying and serializing messages that are
directed to an object in the same PE.

This adds LOCAL_MSG_OPTIM support for groups.

This fix also moves the buffer where these messages are stored to
the destination object, so that when the object migrates, the
messages migrate with it.

Also, the implementation of the pool used for this buffer has
been optimized.

Change-Id: I0f15828d2fa4c49609840c5751ce1b084516de7b