charmpy: General error handling and CkAbort with detailed info
General Python errors/exceptions (either in charmpy or unhandled
errors in user code) are now handled in such a way that the error
information and Python stacktrace is printed and a CkAbort is
triggered.

Previously, general Python errors were not being properly handled.
As such, they would be printed but the program could continue
running, hang or end up CkAborting on an unrelated error.

Change-Id: Ib29fa8a525c4bd82626341a119f378f1b2a9641e