Made charmpy independent of ctypes
Moved ctypes code out of charmpy module. To access the C++ shared
library, charmpy now uses a CharmLib object. Currently there is
only charmlib_ctypes.CharmLib, but the idea is to implement more
efficient methods in the future (e.g. using CFFI, C-extension
module, etc).

Parts of the API have been cleaned or simplified as a result of
this patch.

Change-Id: I348e0e9b64cf74259897ca6d80f1ef9e04cde7f3