charmpy: Direct copy (bypass pickling) of common array types
This feature requires Python 3 and cffi.

Entry method arguments that are bytearrays (bytes), array.array or
numpy.arrays will not be pickled and instead will be directly copied
from the Python object to the Charm++ message with just one copy.

particle example has been updated to use array.array when sending
because it is much faster than sending list of custom types.

Change-Id: Ic3b6de7ec20c0edeb50e1312f121d8ae2678369a