charmpy: Remove use of Singleton
Singleton class was not a full singleton solution as it did not
prevent the constructor from being called multiple times.

More importantly, use of singleton is not needed because a
module's code is only run once even if it is imported multiple
times.

Usage of Readonlies has been changed as a result to take into
account that 'readonlies' object exists in charmpy as soon as the
module is first imported. But legacy API is also maintained for now.

Change-Id: I7c7ae2bad095b2932a59b3274a508637a2c787de