charmpy: Fixed pickling bug with decorated contribute method
Previously, the contribute method of chares was decorated and
modified at runtime if PROFILING=True. The problem is that since
the method is changed after the first definition of the method,
there are pickling errors when the chare migrates.

To avoid this, the method that is modified with the profiling
wrapper is now charm.contribute()

Change-Id: I2b4bdb0abda71be79f3b990913ed1c6eb8f452db