charmpy: Added cython-based layer for increased performance
There can be large performance gains, particularly with
fine-grained entry methods and small msg sizes.

This adds a cython_libcharm layer as an alternative to the current
cffi and ctypes layers, for faster interfacing with libcharm, but
it also cythonizes the packMsg and unpackMsg methods in Charm, and
tightly integrates them with the new cython_libcharm layer.

The cython layer is only supported with Python 3+

auto_test.py has been modified to test all interfaces (cffi,
ctypes and cython). The first two are tested with Py2 and Py3,
cython is tested only with Py3.

Change-Id: I759f42fb9e9bef6e2ec2d47843d4c748b31a2468