charmpy: Added support for creating singleton chares
The API to create singleton chares is:
chare_proxy = Chare(MyChare, args=[1,2,3], onPe=3)

Internally, this is creating an array with only one element using
dynamic insertion. This is likely a temporary way to implement this,
until we make a wrapper for Charm++'s implementation of singleton
chares. Even if we change the underlying implementation later,
the user API won't change.

Change-Id: I1a5e1f4057ffa38610ee91a04f7f4a70fbba949e