Added charm.awaitCreation() to wait for Group and Array creation
This is a convenience function, to allow blocking (inside threaded
entry methods) for the creation of specified Groups and Arrays.
The syntax is:

charm.awaitCreation(*proxies)

For example:

array_proxy = Array(MyChare, 100)
group_proxy = Group(MyChare)
charm.awaitCreation(array_proxy, group_proxy)

NOTE: This only works if the creation calls and the awaitCreation
call happen in the same entry method.

Modified the chares.py tutorial example so that it can exit
automatically after the collections have been created, and added
it to auto_test.py

Change-Id: I8a1d3c7535f1d0233110468658523b33c5e08d44