Allow some methods of charm object to be remotely callable
The charm object is present in every process and represents the
Charm4py runtime. This patch effectively turns it into a chare, by
creating a Group of CharmRemote objects which act as an interface
to the charm object in every process. The new chare can be
accessed via the 'thisProxy' attribute of 'charm'.