Support keyword arguments when calling entry methods
This should now include full support for keyword arguments when
calling entry methods. This includes explicitly naming arguments,
passing them in arbitrary order as long as Python can resolve the
call, and using default values.

Note that best performance is achieved when arguments are not
named when calling entry methods.