particle example: don't use updateGlobals to avoid Charm++ lb bug
Charm++ currently starts load balancing as soon as the mainchare
is done, even if there are no array elements. And when array
elements are created, it will call resumeFromSync on them even if
they never called AtSync. This will be fixed in an upcoming
release of Charm++, but for now the solution is to avoid using
updateGlobals when doing load balancing.