Add distributed exception handling
Exceptions during remote method invocation are sent back to the
caller if using a return mode > 0 (i.e. if a future is requested
when making the remote call). In most cases, the exception will
be raised when doing future.get(). This gives the caller the
option to handle exceptions that are triggered remotely.

This is especially useful in interactive mode, where we want to
give users the option to recover from errors and mistakes
instead of aborting the program.

This works when doing broadcast calls too. If using ret=1, one
exception (from the possibly multiple exceptions) will be
returned via a reduction (it is enough for one of the chares
to raise an exception for this to happen). If using ret=2,
because the caller is expecting a list of return values
(collected via gather), any exceptions raised by callees
will be part of this list of values.

If a return value is not requested when making remote calls and
an exception is raised, it is considered unhandled. In
non-interactive mode, this will abort the program. In
interactive mode, the exception is sent to PE 0 (using a
tree structure where overlapping exceptions in the same
time period are suppressed) and printed.