Add Chare.reduce(callback, data=None, reducer=None) method
This is a new API for doing reductions as an alternative to
Chare.contribute(). Chare.contribute() is maintained for backward
compatibility. Note that reduce takes a callback as the first
argument, because callback is always required. This allows doing
empty reductions by just passing the callback, for example:

self.reduce(callback)