Don't include proxy when pickling Futures
Currently there is no need to include a proxy when pickling
futures. We can get better performance by avoiding it.

Objects with active threads can't migrate, so the location to send
the result of a non-collective future is the PE where it was
created (we pickle the PE number instead).

There is no use case right now for pickling collective futures,
so we don't need to include the proxy for these either.