Fix object wait condition queues not being flushed after future resumed
Previously, the value of a future was deposited from an entry method
of the object to which the future belongs. So the wait queues were
automatically flushed after the resumed thread finishes/pauses.

But after the change that sends futures to CharmRemote, the value
is now deposited from an entry method of CharmRemote, and the wait
queues are not flushed automatically. So we do it manually after
the resumed thread finishes/pauses.