Fix object wait condition queues not being flushed after future resumed
Before 31d07a6, the value of a future was deposited from an entry
method of the object to which the future belongs. So the wait
queues were automatically flushed after the resumed thread
finishes/pauses and the deposit entry method finishes.

But after 31d07a6, futures are sent to CharmRemote, and the value
is now deposited from an entry method of CharmRemote, so
the wait queues are not flushed automatically. So we do it
manually after the resumed thread finishes/pauses.