Use incremental future IDs instead of pool, for better error detection
There are some future-related errors that can be hard to detect.
For example, attempting to reuse futures by sending values to a
future that already received them. If new futures immediately
reuse old future IDs (fid), this can end up sending the value to
a new future, which can lead to weird application errors or
deadlocks.

Instead, this patch gives incremental fids to futures created
on the same PE. When the limit is reached (currently 65535 because
Charm++ normally uses unsigned short for reference numbers carried
in callbacks) it restarts at 1. As such, resending to a used
future will almost certainly trigger a future not found error
instead.

Also, there is no pool so it uses less memory.