Several improvements to charm.pool
- To specify that tasks need to run as coroutines, the user
instead has to decorate these functions with @coro. This makes
it consistent with how it works for chare entry methods. Also,
the 'allow_nested' parameter doesn't need to be passed anymore to
pool.map() and related functions.

- Can spawn individual tasks using charm.pool.Task(func, args)
An individual Task does not implicitly return a result, so they
can spawn other tasks without having to run as a coroutine.
Note that users can pass whatever they want to Tasks, so they
can pass futures that send results somewhere if they want.

- Fixed issue that could potentially lead to weird bugs when
calling the pool async functions. If these are called on PE 0,
the arguments are passed by reference to the PoolScheduler,
and since the functions return immediately, if users modify
the arguments this will cause issues. But this doesn't happen
if the functions are called from PEs other than 0, because
the arguments are immediately packed and sent to the PoolScheduler.
To simplify things and not require users to always copy the
arguments in these special cases, Pool copies them automatically
when called on PE 0.