Add awaitable keyword to proxy entry method calls; new ret behavior
When calling an entry method via a proxy, ret=True means that I
want a return value, regardless of whether it is point-to-point
call or broadcast. In the case of broadcasts, the return value
will be a list of the return values of every call, sorted by element
index (which can be costly).

Use awaitable=True if you just want to wait for completion.

Both awaitable=True and ret=True return a future.